Create ReservedRockets component
