+++ src/routes/Profile.js
@@ -1,23 +1,10 @@
-import React, { useEffect, useState } from 'react';
-import { useSelector } from 'react-redux';
+import React from 'react';
+import ReservedRockets from '../components/ReservedRockets';
 
-const Profile = () => {
-  const rocketsList = useSelector((state) => state.rockets.rocketsList);
-  const [rocketItems, setRocketItems] = useState([]);
-  useEffect(() => {
-    const reservedRockets = rocketsList.filter((rocket) => rocket.reserved === true);
-    setRocketItems(reservedRockets);
-  }, [rocketsList]);
-  return (
-    <div>
-      <h2>My Rockets</h2>
-      <ul>
-        {rocketItems.map((item) => (
-          <li key={item.id}>{item.rocket_name}</li>
-        ))}
-      </ul>
-    </div>
-  );
-};
+const Profile = () => (
+  <div>
+    <ReservedRockets />
+  </div>
+);
 
 export default Profile;
